requirements.txt
faker
black
flake8
pytest
pymongo
jupyter
python-dotenv
google-cloud-firestore
google-cloud-storage
google-cloud-bigquery